agent/requirements.txt
python-binance>=1.0.19
python-dotenv>=1.0.1
anthropic>=0.47.0